refactor(worker): split hook_supervisor.rs and daemon.rs by responsibility
CodeScene's CI gate flags both as Low Cohesion (hook_supervisor.rs: 4
responsibilities across 35 functions; daemon.rs: 4->5). The local `cs
delta` run treats a threshold-crossing count as Degraded rather than New
and doesn't fail, which is why this got missed before merge.

hook_supervisor.rs splits into sibling modules along its actual seams:
- hook_process.rs: spawning and monitoring hook server child processes
- hook_registration.rs: KV registration watches and their bucket bookkeeping
- hook_crash.rs: fail-closed routing (crash markers, rejectors)
- hook_supervisor.rs: the supervisor type and start/spawn orchestration

daemon.rs splits the same way:
- daemon_background.rs: on-demand tool-server/hook/subagent-toolset startup
- daemon_runtime.rs: WorkerRuntime and SessionActivate handling/lease logic
- daemon_turn_input.rs: deriving a turn's input from reconstructed state
- daemon_session_exec.rs: running a claimed session's turn loop to completion
- daemon.rs: bootstrap, subjects, and the run_worker_daemon entrypoint

WorkerRuntime's impl is split across daemon_runtime/daemon_turn_input/
daemon_session_exec via multiple impl blocks on the same struct, matching
how tool_supervisor.rs already reaches into tool_registry.rs's pub(super)
items. No behavior changes; all moved code is copied verbatim.

Updated the #917 structural regression test in tests/nats_worker.rs, which
greps daemon.rs for load_events_latest_async call sites: it now greps the
daemon_turn_input.rs/daemon_session_exec.rs siblings those calls moved into.
