+++ crates/harnx-runtime/src/nats_worker/daemon.rs
@@ -1,24 +1,9 @@
 //! Worker daemon and session activation.
 
-use super::agent_loop::{
-    build_mid_turn_injection_callback, fold_new_user_messages_since,
-    run_agent_loop_with_nats_inner, RunAgentLoopArgs,
-};
-use super::backend::NatsSessionLogBackend;
-use super::control::{control_subject, ControlCommand};
-use super::hook_supervisor::{HookServerStartConfig, HookServerSupervisor};
-use super::server_reconciler::{
-    all_enabled_tool_servers, build_server_reconciler, tool_servers_for_activation,
-    ServerReconciler,
-};
-use super::subagent_toolset::SubagentToolset;
-use super::tool_supervisor::{ToolServerStartConfig, ToolServerSupervisor};
-use crate::config::{
-    list_agents, resolve_local_nats_server_config, server_display_name, Config, GlobalConfig,
-    Input, ToolServerConfig,
-};
-use crate::nats_lease::{NatsLeaseAcquireParams, NatsLeaseConfig, NatsSessionLease};
-use crate::nats_metrics;
+use super::daemon_background::launch_worker_services;
+use super::daemon_runtime::WorkerRuntime;
+use crate::config::GlobalConfig;
+use crate::nats_lease::{NatsLeaseConfig, NatsSessionLease};
 use crate::nats_session_index;
 use anyhow::{Context, Result};
 use async_nats::header::{HeaderValue, NATS_MESSAGE_ID};
@@ -31,12 +16,11 @@ use chrono::Utc;
 use futures_util::StreamExt;
 use harnx_core::instance::{ServerScope, HARNX_SERVER_SCOPE};
 use serde::{Deserialize, Serialize};
-use std::collections::{HashMap, HashSet};
-use std::sync::atomic::{AtomicU64, Ordering};
+use std::collections::HashMap;
+use std::sync::atomic::AtomicU64;
 use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::Mutex;
-use tokio::task::JoinHandle;
 
 // ---------------------------------------------------------------------------
 // P2.1 worker daemon + SessionActivate dispatch
@@ -46,21 +30,6 @@ const WORK_NOTIFY_STREAM_PREFIX: &str = "WORK_NOTIFY_";
 const WORK_NOTIFY_CONSUMER_PREFIX: &str = "worker-";
 const WORK_NOTIFY_ACK_WAIT: Duration = Duration::from_secs(30);
 const WORK_NOTIFY_INACTIVE_THRESHOLD: Duration = Duration::from_secs(60 * 60);
-/// Longest a session start will wait for the first tool-server registration
-/// round. Above the round's own budget, so it only fires if that round wedges.
-const INITIAL_TOOL_REGISTRATION_WAIT: Duration = Duration::from_secs(30);
-/// Longest `handle_activation` waits for this session's own tool servers to
-/// start before giving up and continuing anyway.
-///
-/// Deliberately well under `WORK_NOTIFY_ACK_WAIT`: tool-server starts now run
-/// concurrently (see `ServerReconciler::session_started`), so this only needs
-/// to cover one server's own startup timeout plus margin, not the sum of
-/// several. Comfortably clearing this margin matters because JetStream
-/// redelivers an unacked activation at `WORK_NOTIFY_ACK_WAIT` with
-/// `max_deliver: -1` — a session that never manages to ack loops forever,
-/// each redelivery re-acquiring the lease and fencing the still-running
-/// previous attempt.
-const SESSION_TOOL_SERVER_START_TIMEOUT: Duration = Duration::from_secs(20);
 
 /// Configuration for a worker daemon instance.
 #[derive(Debug, Clone)]
@@ -240,24 +209,14 @@ pub async fn publish_session_activate(
     Ok(ack.sequence)
 }
 
-/// Shared, borrowed context for the `derive_*_turn_input` helpers. Groups the
-/// three references they all thread through so each helper stays within the
-/// function-argument budget.
-#[derive(Clone, Copy)]
-struct TurnInputCtx<'a> {
-    activation: &'a SessionActivate,
-    per_session: &'a GlobalConfig,
-    backend: &'a NatsSessionLogBackend,
-}
-
-struct WorkerStartup {
-    jetstream: jetstream::Context,
-    client: async_nats::Client,
-    consumer: jetstream::consumer::Consumer<pull::Config>,
+pub(super) struct WorkerStartup {
+    pub(super) jetstream: jetstream::Context,
+    pub(super) client: async_nats::Client,
+    pub(super) consumer: jetstream::consumer::Consumer<pull::Config>,
     /// JetStream replica count for `daemon.cluster`, resolved once here so
     /// every bucket this worker creates (leases, session index, tool/hook
     /// registries) agrees on it instead of drifting to a per-site default.
-    replicas: usize,
+    pub(super) replicas: usize,
 }
 
 async fn prepare_worker_startup(
@@ -297,7 +256,7 @@ async fn prepare_worker_startup(
     })
 }
 
-fn spawn_readiness_publisher(client: async_nats::Client, daemon: &WorkerDaemonConfig) {
+pub(super) fn spawn_readiness_publisher(client: async_nats::Client, daemon: &WorkerDaemonConfig) {
     let subject = worker_ready_subject(&daemon.cluster);
     let worker_id = daemon.worker_id.clone();
     tokio::spawn(async move {
@@ -318,7 +277,7 @@ fn spawn_readiness_publisher(client: async_nats::Client, daemon: &WorkerDaemonCo
     });
 }
 
-async fn optional_session_index(
+pub(super) async fn optional_session_index(
     jetstream: &jetstream::Context,
     replicas: usize,
 ) -> Option<async_nats::jetstream::kv::Store> {
@@ -332,270 +291,6 @@ async fn optional_session_index(
             None
         }
     }
-}
-
-pub(crate) fn tool_servers_matching_use_tools(
-    servers: &[ToolServerConfig],
-    agent_package: Option<&str>,
-    namespaced_use_tools: &[String],
-) -> Vec<ToolServerConfig> {
-    let mut seen_names = HashSet::new();
-    servers
-        .iter()
-        .filter(|server| {
-            if !server.enabled {
-                return false;
-            }
-            let display_name =
-                server_display_name(&server.name, server.package.as_deref(), agent_package);
-            let matches = namespaced_use_tools.iter().any(|selector| {
-                super::super::config::selector_could_match_server(selector, &display_name)
-            });
-            let identity_token = harnx_toolset::server_identity_token(
-                server.package.as_deref(),
-                &server.name,
-                &server.name,
-            );
-            matches && seen_names.insert(identity_token)
-        })
-        .cloned()
-        .collect()
-}
-
-/// Whether this worker should launch its own tool servers.
-///
-/// A worker with nothing configured to spawn must not resolve a broker
-/// address, because resolving falls back to starting a shared NATS server.
-/// Mirrors `start_global_hooks`'s `hooks.entries.is_empty()` check. Pulled out
-/// of `start_local_tool_servers` so the decision itself — not a copy of it —
-/// is what tests exercise.
-pub(super) fn should_start_tool_servers(
-    manage_servers: bool,
-    servers: &[ToolServerConfig],
-) -> bool {
-    manage_servers && !servers.is_empty()
-}
-
-async fn start_local_tool_servers(
-    daemon: &WorkerDaemonConfig,
-    client: async_nats::Client,
-    instance_id: &harnx_core::instance::ServerScope,
-    servers: &[crate::config::ToolServerConfig],
-) -> Option<ToolServerSupervisor> {
-    if !should_start_tool_servers(daemon.manage_servers, servers) {
-        return None;
-    }
-    let result = async {
-        let server = resolve_local_nats_server_config().await?;
-        let token = server
-            .token
-            .as_deref()
-            .context("local NATS tool servers require HARNX_NATS_TOKEN")?;
-        let start = ToolServerStartConfig::new(client, instance_id.clone(), &server.url, token)
-            .with_replicas(server.replicas)
-            .with_tls(&harnx_nats_common::connect::NatsEndpoint::from(&server));
-        ToolServerSupervisor::start_local(start, servers)
-            .await
-            .context("start local NATS tool servers")
-    }
-    .await;
-    optional_tool_server(result)
-}
-
-/// Test entry point for [`start_local_tool_servers`].
-///
-/// Public for regression coverage that a consuming worker (the default) never
-/// spawns child tool servers, and that a managing worker with nothing
-/// configured spawns none either. Not part of the crate's real API —
-/// production code always goes through [`start_local_tool_servers`] with an
-/// already connected client obtained during worker startup.
-///
-/// Deliberately does none of `start_local_tool_servers`'s own gating here: it
-/// builds a client and calls straight through, so a regression in the real
-/// function's guard shows up as this shim actually starting a tool-server
-/// supervisor (or, without a reachable broker, hanging/erroring on the
-/// connect) instead of silently returning `None` from a copy of the check.
-#[doc(hidden)]
-pub async fn start_local_tool_servers_for_test(
-    daemon: &WorkerDaemonConfig,
-    config: &GlobalConfig,
-) -> Option<ToolServerSupervisor> {
-    let (servers, _hooks) = configured_worker_services(config);
-    let server = resolve_local_nats_server_config().await.ok()?;
-    let token = server.token.as_deref()?;
-    let client = async_nats::ConnectOptions::new()
-        .token(token.to_string())
-        .connect(&server.url)
-        .await
-        .ok()?;
-    let instance_id = harnx_core::instance::ServerScope::new();
-    start_local_tool_servers(daemon, client, &instance_id, &servers).await
-}
-
-async fn start_global_hooks(
-    daemon: &WorkerDaemonConfig,
-    client: async_nats::Client,
-    instance_id: &harnx_core::instance::ServerScope,
-    hooks: &harnx_core::hooks::HooksConfig,
-) -> Option<HookServerSupervisor> {
-    if !daemon.manage_servers || hooks.entries.is_empty() {
-        return None;
-    }
-    let result = async {
-        let server = resolve_local_nats_server_config().await?;
-        let token = server
-            .token
-            .as_deref()
-            .context("local NATS hook servers require HARNX_NATS_TOKEN")?;
-        let start = HookServerStartConfig::new(client, instance_id.clone(), &server.url, token)
-            .with_replicas(server.replicas)
-            .with_tls(&harnx_nats_common::connect::NatsEndpoint::from(&server));
-        HookServerSupervisor::start_local(start, hooks, "global")
-            .await
-            .context("start global NATS hook servers")
-    }
-    .await;
-    match result {
-        Ok(supervisor) => Some(supervisor),
-        Err(error) => {
-            // Failures happen before the supervisor can own cleanup or while its KV
-            // route is unavailable. Publishing here would either reuse the failed
-            // route or leave an unowned rejector behind after worker shutdown. Keep
-            // the worker available; unreadable registries fail closed at discovery.
-            log::warn!("global NATS hook servers disabled: {error:#}");
-            None
-        }
-    }
-}
-
-fn optional_tool_server<T>(result: Result<T>) -> Option<T> {
-    match result {
-        Ok(supervisor) => Some(supervisor),
-        Err(error) => {
-            log::warn!("local NATS tool servers disabled; continuing with stdio tools: {error:#}");
-            None
-        }
-    }
-}
-
-/// Everything [`start_subagent_toolset`] needs, bundled so adding `replicas`
-/// didn't push it to a 6th bare argument.
-struct SubagentToolsetStart {
-    agent: String,
-    cluster: String,
-    instance_id: harnx_core::instance::ServerScope,
-    client: async_nats::Client,
-    jetstream: jetstream::Context,
-    replicas: usize,
-}
-
-async fn start_subagent_toolset(start: SubagentToolsetStart) -> Result<JoinHandle<Result<()>>> {
-    let SubagentToolsetStart {
-        agent,
-        cluster,
-        instance_id,
-        client,
-        jetstream,
-        replicas,
-    } = start;
-    let registration_context = jetstream.clone();
-    let toolset = Arc::new(SubagentToolset::new(
-        agent,
-        cluster,
-        client.clone(),
-        jetstream,
-    ));
-    let server_name = harnx_toolset::Toolset::name(toolset.as_ref()).to_string();
-    let identity_token = harnx_toolset::server_identity_token(None, "", &server_name);
-    let registration_key = harnx_toolset_server::registration_key(&instance_id, &identity_token);
-    let connection = harnx_nats_common::connect::NatsConnection { client, replicas };
-    let server = tokio::spawn(harnx_toolset_server::serve_with_client(
-        toolset,
-        instance_id,
-        connection,
-    ));
-
-    let registration = tokio::time::timeout(Duration::from_secs(5), async {
-        loop {
-            if server.is_finished() {
-                anyhow::bail!("sub-agent tool server exited before registering");
-            }
-            match registration_context
-                .get_key_value(harnx_toolset_server::TOOL_REGISTRY_BUCKET)
-                .await
-            {
-                Ok(registry) if registry.get(&registration_key).await?.is_some() => break,
-                Ok(_) | Err(_) => tokio::time::sleep(Duration::from_millis(25)).await,
-            }
-        }
-        Result::<()>::Ok(())
-    })
-    .await;
-
-    match registration {
-        Ok(Ok(())) => Ok(server),
-        Ok(Err(error)) => {
-            server.abort();
-            Err(error)
-        }
-        Err(_) => {
-            server.abort();
-            anyhow::bail!("sub-agent tool server did not register within 5s")
-        }
-    }
-}
-
-struct WorkerServices {
-    background: Arc<Mutex<Option<BackgroundServices>>>,
-    session_index: Option<async_nats::jetstream::kv::Store>,
-    tools_attempted: tokio::sync::watch::Receiver<bool>,
-    /// `None` for a consuming worker, or a managing worker with nothing
-    /// configured to spawn. Tool servers now start on demand per session
-    /// (see `handle_activation`) rather than as one batch here.
-    server_reconciler: Option<Arc<ServerReconciler>>,
-}
-
-/// Hooks and sub-agent toolsets, started after the worker has already
-/// announced readiness. Held only to keep the children alive for the
-/// worker's lifetime. Tool servers are no longer part of this: they start
-/// per session through `WorkerRuntime::server_reconciler`.
-struct BackgroundServices {
-    _global_hook_supervisor: Option<HookServerSupervisor>,
-    _subagent_tool_servers: Vec<JoinHandle<Result<()>>>,
-}
-
-pub(super) fn configured_worker_services(
-    config: &GlobalConfig,
-) -> (Vec<ToolServerConfig>, harnx_core::hooks::HooksConfig) {
-    let config = config.read();
-    let agent_package = config
-        .agent
-        .as_ref()
-        .and_then(|agent| harnx_core::package_namespace::pkg_from_qualified(agent.name()));
-    let mut selectors = Vec::new();
-    let use_tools = config
-        .agent
-        .as_ref()
-        .and_then(|agent| agent.use_tools())
-        // Fall back when no agent is active or the active agent has no use_tools.
-        .or_else(|| config.use_tools.clone())
-        .unwrap_or_default();
-    for selector in use_tools {
-        if let Some(namespaced) = namespaced_selector(&selector, agent_package) {
-            selectors.push(namespaced);
-        }
-        selectors.push(selector);
-    }
-    let servers = tool_servers_matching_use_tools(&config.tool_servers, agent_package, &selectors);
-    (servers, config.hooks.clone().unwrap_or_default())
-}
-
-fn namespaced_selector(selector: &str, package: Option<&str>) -> Option<String> {
-    if selector == "*" {
-        return None;
-    }
-    let namespaced = harnx_core::package_namespace::namespace_use_tools_entry(package?, selector);
-    (namespaced != selector).then_some(namespaced)
 }
 
 /// Install the activation's agent into the per-session config.
@@ -620,7 +315,7 @@ pub(super) fn install_activation_agent(per_session: &GlobalConfig, agent_name: &
     let mut agent = match retrieved {
         Ok(agent) => agent,
         // No file and no built-in by this name: nothing to load.
-        Err(error) if !Config::agent_file(agent_name).exists() => {
+        Err(error) if !crate::config::Config::agent_file(agent_name).exists() => {
             log::warn!(
                 "activation agent '{agent_name}' not available, using worker config: {error:#}"
             );
@@ -635,146 +330,6 @@ pub(super) fn install_activation_agent(per_session: &GlobalConfig, agent_name: &
         .with_context(|| format!("activate agent '{agent_name}'"))?;
     cfg.require_agent_shared_variables()
         .with_context(|| format!("initialize variables for agent '{agent_name}'"))
-}
-
-async fn launch_worker_services(
-    config: &GlobalConfig,
-    daemon: &WorkerDaemonConfig,
-    startup: &WorkerStartup,
-    instance_id: &harnx_core::instance::ServerScope,
-) -> Result<WorkerServices> {
-    // Announce readiness before starting hooks and sub-agent toolsets. Those
-    // can take tens of seconds — or never finish, when a server is
-    // misconfigured — and neither is required for the worker to accept and
-    // run a session. Gating readiness on them made a single broken server
-    // stall the front-end past its startup deadline, which turned an
-    // intentionally non-fatal degradation into an unusable CLI. Tool servers
-    // no longer start here at all: each session's own servers start (and are
-    // waited on) in `handle_activation` when that session activates.
-    spawn_readiness_publisher(startup.client.clone(), daemon);
-
-    let background = Arc::new(Mutex::new(None));
-    // There is no longer a global tool-server registration round for a
-    // session to wait on: `handle_activation` waits on the servers it just
-    // asked the reconciler to start instead. Start this already-settled so
-    // `await_initial_tool_registration` is a no-op, kept for the hooks/
-    // sub-agent-toolset background task shape rather than removed outright.
-    let (_tools_attempted_tx, tools_attempted) = tokio::sync::watch::channel(true);
-
-    let all_tool_servers = all_enabled_tool_servers(config);
-    let server_reconciler = build_server_reconciler(
-        daemon,
-        startup.client.clone(),
-        instance_id,
-        &all_tool_servers,
-    )
-    .await;
-
-    spawn_background_services(BackgroundServicesCtx {
-        config: config.clone(),
-        daemon: daemon.clone(),
-        client: startup.client.clone(),
-        jetstream: startup.jetstream.clone(),
-        instance_id: instance_id.clone(),
-        slot: Arc::clone(&background),
-        replicas: startup.replicas,
-    });
-
-    let session_index = optional_session_index(&startup.jetstream, startup.replicas).await;
-    Ok(WorkerServices {
-        background,
-        session_index,
-        tools_attempted,
-        server_reconciler,
-    })
-}
-
-/// Historically blocked until the first tool-server registration round had
-/// finished, so a session's registry snapshot included whatever managed to
-/// come up. Tool servers now start per session (`handle_activation` awaits
-/// `ServerReconciler::session_started` directly), so `tools_attempted` is
-/// already settled by the time this runs and it returns immediately. Kept
-/// rather than removed so a future global round has somewhere to plug back
-/// in without re-threading `handle_activation`.
-async fn await_initial_tool_registration(tools_attempted: &tokio::sync::watch::Receiver<bool>) {
-    if *tools_attempted.borrow() {
-        return;
-    }
-    let mut attempted = tools_attempted.clone();
-    log::debug!("waiting for the first tool-server registration round");
-    // The only sender lives in the background task; if it is gone the round can
-    // never complete and there is nothing left to wait for. Cap the wait so a
-    // wedged round costs the session its tools rather than the whole turn — the
-    // round is bounded by the per-server startup timeout, so reaching this cap
-    // means something below it is stuck.
-    if tokio::time::timeout(
-        INITIAL_TOOL_REGISTRATION_WAIT,
-        attempted.wait_for(|done| *done),
-    )
-    .await
-    .is_err()
-    {
-        log::warn!(
-            "tool servers did not finish their first registration round within {}s; \
-             starting this session with the tools registered so far",
-            INITIAL_TOOL_REGISTRATION_WAIT.as_secs()
-        );
-    }
-}
-
-/// Owned inputs for the post-readiness startup task.
-struct BackgroundServicesCtx {
-    config: GlobalConfig,
-    daemon: WorkerDaemonConfig,
-    client: async_nats::Client,
-    jetstream: jetstream::Context,
-    instance_id: harnx_core::instance::ServerScope,
-    slot: Arc<Mutex<Option<BackgroundServices>>>,
-    replicas: usize,
-}
-
-fn spawn_background_services(ctx: BackgroundServicesCtx) {
-    tokio::spawn(async move {
-        let BackgroundServicesCtx {
-            config,
-            daemon,
-            client,
-            jetstream,
-            instance_id,
-            slot,
-            replicas,
-        } = ctx;
-        // Tool servers aren't started here anymore — each session's own
-        // servers start on demand through `WorkerRuntime::server_reconciler`.
-        let (_worker_tool_servers, global_hooks) = configured_worker_services(&config);
-        let global_hook_supervisor =
-            start_global_hooks(&daemon, client.clone(), &instance_id, &global_hooks).await;
-
-        let mut subagent_tool_servers = Vec::new();
-        for agent in list_agents() {
-            match start_subagent_toolset(SubagentToolsetStart {
-                agent: agent.clone(),
-                cluster: daemon.cluster.clone(),
-                instance_id: instance_id.clone(),
-                client: client.clone(),
-                jetstream: jetstream.clone(),
-                replicas,
-            })
-            .await
-            {
-                Ok(handle) => subagent_tool_servers.push(handle),
-                // One unusable sub-agent must not cost the others their toolset.
-                Err(error) => {
-                    log::warn!("sub-agent toolset for '{agent}' unavailable: {error:#}")
-                }
-            }
-        }
-
-        *slot.lock().await = Some(BackgroundServices {
-            _global_hook_supervisor: global_hook_supervisor,
-            _subagent_tool_servers: subagent_tool_servers,
-        });
-    });
 }
 
 /// Run a worker daemon: pull `SessionActivate` notifications, claim each via a
@@ -823,893 +378,4 @@ pub async fn run_worker_daemon(
         }
     }
     Ok(())
-}
-
-/// Borrowed parameters for [`WorkerRuntime::spawn_control_listener`].
-struct ControlListenerCtx<'a> {
-    client: &'a async_nats::Client,
-    session_id: &'a str,
-    lease: &'a Arc<NatsSessionLease>,
-    backend: &'a NatsSessionLogBackend,
-    abort_signal: &'a crate::utils::AbortSignal,
-}
-
-/// Borrowed parameters for [`WorkerRuntime::prepare_and_ack_activation`].
-struct ActivationAckCtx<'a> {
-    activation: &'a SessionActivate,
-    message: &'a async_nats::jetstream::Message,
-    lease: &'a Arc<NatsSessionLease>,
-    abort_signal: &'a crate::utils::AbortSignal,
-}
-
-struct WorkerRuntime {
-    config: GlobalConfig,
-    instance_id: harnx_core::instance::ServerScope,
-    _background_services: Arc<Mutex<Option<BackgroundServices>>>,
-    tools_attempted: tokio::sync::watch::Receiver<bool>,
-    /// `None` for a consuming worker, or a managing worker with nothing
-    /// configured to spawn anywhere.
-    server_reconciler: Option<Arc<ServerReconciler>>,
-    #[allow(dead_code)]
-    cluster: String,
-    manage_servers: bool,
-    worker_id: String,
-    lease: NatsLeaseConfig,
-    jetstream: jetstream::Context,
-    session_index: Option<async_nats::jetstream::kv::Store>,
-    /// Shared NATS client for control-plane subscriptions (cloned per session
-    /// rather than reconnecting on each activation).
-    client: async_nats::Client,
-    call_fn: Option<crate::agent_loop::AgentCallFn>,
-    generation: AtomicU64,
-    active: Mutex<HashMap<String, JoinHandle<()>>>,
-}
-
-impl WorkerRuntime {
-    async fn already_running(&self, session_id: &str) -> bool {
-        let mut active = self.active.lock().await;
-        active.retain(|_, handle| !handle.is_finished());
-        active.contains_key(session_id)
-    }
-
-    /// A worker that doesn't manage its own servers, or has nothing
-    /// configured to spawn anywhere, has no reconciler and this is a no-op.
-    ///
-    /// Bounded by `SESSION_TOOL_SERVER_START_TIMEOUT`, comfortably under the
-    /// activation ack window: a session degraded to fewer tools is far better
-    /// than an unacked activation that JetStream redelivers forever. On
-    /// timeout the reconciler task is left running rather than aborted — it
-    /// still finishes starting (or gives up on) each server and updates
-    /// `ServerReconciler`'s own bookkeeping correctly; this call just stops
-    /// waiting for it.
-    async fn start_session_tool_servers(&self, activation: &SessionActivate) {
-        let Some(reconciler) = self.server_reconciler.clone() else {
-            return;
-        };
-        let servers = tool_servers_for_activation(&self.config, &activation.agent);
-        if servers.is_empty() {
-            return;
-        }
-        let server_names = servers
-            .iter()
-            .map(|server| server.name.clone())
-            .collect::<Vec<_>>()
-            .join(", ");
-        let session_id = activation.session_id.clone();
-        let task = tokio::spawn(async move {
-            reconciler.session_started(&session_id, servers).await;
-        });
-        match tokio::time::timeout(SESSION_TOOL_SERVER_START_TIMEOUT, task).await {
-            Ok(Ok(())) => {}
-            Ok(Err(join_error)) => {
-                log::warn!(
-                    "session '{}' tool-server startup task panicked: {join_error}",
-                    activation.session_id
-                );
-            }
-            Err(_) => {
-                log::warn!(
-                    "session '{}' tool-server startup ({}) exceeded {}s; continuing this \
-                     activation without waiting further (still starting in the background)",
-                    activation.session_id,
-                    server_names,
-                    SESSION_TOOL_SERVER_START_TIMEOUT.as_secs()
-                );
-            }
-        }
-    }
-
-    async fn end_session_tool_servers(&self, session_id: &str) {
-        if let Some(reconciler) = &self.server_reconciler {
-            reconciler.session_ended(session_id).await;
-        }
-    }
-
-    async fn acquire_activation_lease(
-        &self,
-        activation: &SessionActivate,
-    ) -> Result<Option<Arc<NatsSessionLease>>> {
-        let generation = self.generation.fetch_add(1, Ordering::SeqCst);
-        let lease = NatsSessionLease::acquire(NatsLeaseAcquireParams {
-            jetstream: self.jetstream.clone(),
-            session_id: &activation.session_id,
-            worker_id: self.worker_id.clone(),
-            generation,
-            config: self.lease.clone(),
-            session_index: self.session_index.clone(),
-        })
-        .await?;
-        Ok(lease.map(Arc::new))
-    }
-
-    async fn prepare_activation_control(
-        &self,
-        activation: &SessionActivate,
-        lease: &Arc<NatsSessionLease>,
-        abort_signal: &crate::utils::AbortSignal,
-    ) -> Result<JoinHandle<()>> {
-        let backend = NatsSessionLogBackend::new(self.jetstream.clone(), &activation.session_id);
-        match Self::spawn_control_listener(ControlListenerCtx {
-            client: &self.client,
-            session_id: &activation.session_id,
-            lease,
-            backend: &backend,
-            abort_signal,
-        })
-        .await
-        {
-            Ok(task) => Ok(task),
-            Err(error) => {
-                let _ = lease.release().await;
-                Err(error)
-            }
-        }
-    }
-
-    /// Subscribe control and acknowledge the activation, in that order.
-    /// `start_session_tool_servers` already registered this session as a tool-
-    /// server user before either step ran; no task exists yet to release that
-    /// on completion (that happens inside the spawned task `handle_activation`
-    /// creates once this returns `Ok`), so either failure here must release it
-    /// itself or the server stays pinned running for the rest of the worker's
-    /// lifetime.
-    async fn prepare_and_ack_activation(
-        &self,
-        ctx: ActivationAckCtx<'_>,
-    ) -> Result<JoinHandle<()>> {
-        let control_task = match self
-            .prepare_activation_control(ctx.activation, ctx.lease, ctx.abort_signal)
-            .await
-        {
-            Ok(task) => task,
-            Err(error) => {
-                self.end_session_tool_servers(&ctx.activation.session_id)
-                    .await;
-                return Err(error);
-            }
-        };
-        if let Err(error) = ctx.message.ack().await {
-            control_task.abort();
-            let _ = ctx.lease.release().await;
-            self.end_session_tool_servers(&ctx.activation.session_id)
-                .await;
-            return Err(anyhow::anyhow!("ack SessionActivate: {error}"));
-        }
-        Ok(control_task)
-    }
-
-    async fn handle_activation(
-        self: &Arc<Self>,
-        message: async_nats::jetstream::Message,
-    ) -> Result<()> {
-        let activation: SessionActivate =
-            serde_json::from_slice(&message.payload).context("decode SessionActivate")?;
-
-        // Re-activation of an already-running session is a no-op.
-        if self.already_running(&activation.session_id).await {
-            let _ = message.ack().await;
-            return Ok(());
-        }
-
-        // A lease loser leaves the message unacked for its current holder.
-        let Some(lease) = self.acquire_activation_lease(&activation).await? else {
-            return Ok(());
-        };
-
-        // Start (or reuse) this session's own agent's tool servers before the
-        // registry snapshot below is taken, so it sees them.
-        self.start_session_tool_servers(&activation).await;
-
-        // The session snapshots the tool registry below, so give the first
-        // registration round a chance to finish before that snapshot is taken.
-        await_initial_tool_registration(&self.tools_attempted).await;
-        log::info!(
-            "session activate claimed: session_id={} worker_id={} revision={} epoch={}",
-            activation.session_id,
-            lease.worker_id(),
-            lease.fence_token(),
-            activation.epoch
-        );
-
-        let abort_signal = crate::utils::create_abort_signal();
-        // Core-NATS control must be subscribed before activation is
-        // acknowledged; either failing releases the lease and this session's
-        // tool-server refcount (see `prepare_and_ack_activation`).
-        let control_task = self
-            .prepare_and_ack_activation(ActivationAckCtx {
-                activation: &activation,
-                message: &message,
-                lease: &lease,
-                abort_signal: &abort_signal,
-            })
-            .await?;
-        let worker = Arc::clone(self);
-        let session_id = activation.session_id.clone();
-        let task_session_id = session_id.clone();
-        nats_metrics::active_session_started();
-        let handle = tokio::spawn(async move {
-            let snapshot = nats_metrics::snapshot();
-            log::info!(
-                "active session started: session_id={} worker_id={} revision={} active_sessions_per_worker={}",
-                activation.session_id,
-                lease.worker_id(),
-                lease.fence_token(),
-                snapshot.active_sessions_per_worker
-            );
-            let result = worker
-                .execute_session(activation, Arc::clone(&lease), abort_signal, control_task)
-                .await;
-            // `active` is only pruned lazily (see `already_running`'s
-            // `retain`), so this is the one place a finished session's tool
-            // servers can be released — do it before the metrics/log lines
-            // below, which already mark the session as done.
-            worker.end_session_tool_servers(&task_session_id).await;
-            nats_metrics::active_session_finished();
-            let snapshot = nats_metrics::snapshot();
-            log::info!(
-                "active session finished: session_id={} worker_id={} revision={} active_sessions_per_worker={}",
-                task_session_id,
-                lease.worker_id(),
-                lease.fence_token(),
-                snapshot.active_sessions_per_worker
-            );
-            if let Err(error) = result {
-                log::warn!("worker session execution failed: {error:#}");
-            }
-        });
-        self.active.lock().await.insert(session_id, handle);
-        Ok(())
-    }
-
-    async fn derive_continuation_turn_input(
-        &self,
-        ctx: TurnInputCtx<'_>,
-        high_water: Option<u64>,
-    ) -> Option<(Input, Option<u64>)> {
-        // Continuation turns (high_water set) derive input CURSOR-based, not
-        // barrier-based. A user message that arrives DURING a turn is logged at
-        // a seq BELOW that turn's assistant barrier, so the barrier-based
-        // reconstruct would treat it as already-answered and never fold it. The
-        // cursor (high-water mark of messages already fed this activation) is the
-        // authoritative "unanswered" boundary and matches the drain decision.
-        let hw = high_water?;
-        let tail = match ctx.backend.load_events_latest_async().await {
-            Ok(entries) => entries,
-            Err(err) => {
-                log::warn!(
-                    "failed to load session log for continuation drain: session_id={} worker_id={} err={err}",
-                    ctx.backend.session_id(),
-                    self.worker_id,
-                );
-                Vec::new()
-            }
-        };
-        let (new_messages, latest_seq) = fold_new_user_messages_since(&tail, Some(hw));
-        if new_messages.is_empty() {
-            return None;
-        }
-        let (mut input, seed) = self
-            .derive_idle_turn_input(ctx.activation, ctx.per_session, new_messages)
-            .await;
-        input.with_session = true;
-        input.skip_user_log_append = true;
-        Some((input, seed.or(latest_seq)))
-    }
-
-    /// Derive the turn input for an activation from the reconstructed session
-    /// state, honoring cancel tombstones, resumable turns, and pending messages.
-    async fn derive_turn_input(
-        &self,
-        activation: &SessionActivate,
-        lease: &NatsSessionLease,
-        per_session: &GlobalConfig,
-        backend: &NatsSessionLogBackend,
-        _session_index: Option<&async_nats::jetstream::kv::Store>,
-        high_water: Option<u64>,
-    ) -> (Input, Option<u64>) {
-        let ctx = TurnInputCtx {
-            activation,
-            per_session,
-            backend,
-        };
-        if let Some(continuation) = self.derive_continuation_turn_input(ctx, high_water).await {
-            return continuation;
-        }
-
-        let reconstructed = self.reconstruct_session_state(backend).await;
-        log::debug!(
-            "derive_turn_input: session_id={} turn_status={:?} next_turn_count={} resumable_ctx={}",
-            activation.session_id,
-            reconstructed.turn_status,
-            reconstructed.next_turn_messages.len(),
-            reconstructed.resumable_ctx.is_some(),
-        );
-        let (mut input, seed_cursor) = match reconstructed.turn_status {
-            harnx_core::session_reconstruct::TurnStatus::InFlightCancelled => {
-                // Terminal: Cancel tombstone prevents resume. Do NOT consume pending.
-                // The turn is idle; wait for new user input or activation.
-                log::info!(
-                    "session has cancelled turn tombstone; not resuming (session_id={})",
-                    activation.session_id
-                );
-                (crate::config::input::from_str(per_session, "", None), None)
-            }
-            harnx_core::session_reconstruct::TurnStatus::InFlightResumable => {
-                // Resume existing turn (orphan repair handled inside run_agent_loop_with_nats_inner).
-                log::info!(
-                    "resume state: session_id={} worker_id={} revision={} mode=resumable",
-                    activation.session_id,
-                    lease.worker_id(),
-                    lease.fence_token()
-                );
-                // Extract the last_user Message to get both its text (for Input) and log_seq (for cursor).
-                let last_user_msg = reconstructed
-                    .resumable_ctx
-                    .as_ref()
-                    .and_then(|ctx| ctx.last_user.as_ref());
-                let input = if let Some(last_user) = last_user_msg {
-                    crate::config::input::from_str(per_session, &last_user.content.to_text(), None)
-                } else {
-                    crate::config::input::from_str(per_session, "", None)
-                };
-                // Cursor: the log_seq of the last user message that kicked off this resumable turn.
-                // Any messages appended AFTER this (seq > cursor) will be folded mid-turn.
-                let seed_cursor = last_user_msg
-                    .and_then(|msg| msg.log_seq.and_then(|seq| u64::try_from(seq).ok()));
-                (input, seed_cursor)
-            }
-            harnx_core::session_reconstruct::TurnStatus::Idle => {
-                let msg_count = reconstructed.next_turn_messages.len();
-                let result = self
-                    .derive_idle_turn_input(
-                        activation,
-                        per_session,
-                        reconstructed.next_turn_messages,
-                    )
-                    .await;
-                // DEBUG: log the seed_cursor for this path
-                log::debug!(
-                    "derive_turn_input idle: session_id={} seed_cursor={:?} messages_count={}",
-                    activation.session_id,
-                    result.1,
-                    msg_count,
-                );
-                result
-            }
-        };
-        // The NATS worker ALWAYS operates on a session. The input is derived
-        // before `run_agent_loop_with_nats_inner` attaches the session to the
-        // per-session config, so `from_str` sees no session and leaves
-        // `with_session=false`. That would make `save_message` a no-op and the
-        // turn's assistant barrier would never be persisted. Force it true.
-        input.with_session = true;
-        // The folded user messages are ALREADY durable in the log (clients
-        // append them directly) and loaded into `session.messages`. The worker
-        // must not re-append them: doing so duplicates the user message and
-        // reorders the assistant barrier past concurrently-arrived messages,
-        // burying them so they are never folded into a continuation turn.
-        input.skip_user_log_append = true;
-        (input, seed_cursor)
-    }
-
-    /// Idle-state input: fold queued next-turn messages in log order.
-    async fn derive_idle_turn_input(
-        &self,
-        _activation: &SessionActivate,
-        per_session: &GlobalConfig,
-        next_turn_messages: Vec<harnx_core::message::Message>,
-    ) -> (Input, Option<u64>) {
-        let seed_cursor = next_turn_messages
-            .last()
-            .and_then(|message| message.log_seq.and_then(|seq| u64::try_from(seq).ok()));
-        let folded = next_turn_messages
-            .into_iter()
-            .map(|message| message.content.to_text())
-            .filter(|text| !text.is_empty())
-            .collect::<Vec<_>>()
-            .join("\n");
-        (
-            crate::config::input::from_str(per_session, &folded, None),
-            seed_cursor,
-        )
-    }
-
-    async fn execute_session(
-        &self,
-        activation: SessionActivate,
-        lease: Arc<NatsSessionLease>,
-        abort_signal: crate::utils::AbortSignal,
-        control_task: JoinHandle<()>,
-    ) -> Result<()> {
-        // Per-session config clone with the requested agent (loaded from the
-        // worker's OWN config) and the session selected.
-        let per_session = {
-            let base = self.config.read().clone();
-            Arc::new(parking_lot::RwLock::new(base))
-        };
-        let agent_setup = install_activation_agent(&per_session, &activation.agent);
-
-        // Create event sink for live fan-out. `new` seeds `after_seq` from stream once.
-        let event_sink = crate::nats_event_sink::NatsEventSink::new(
-            self.client.clone(),
-            self.jetstream.clone(),
-            activation.session_id.clone(),
-        )
-        .await;
-        let after_seq_observer = event_sink.after_seq_handle();
-        let event_sink = Arc::new(event_sink);
-
-        // Build the backend for control-plane operations and state reconstruction.
-        // Share the `after_seq` high-water mark for event-sink fan-out advisories;
-        // worker tail reads themselves use leader-authoritative `load_events_latest_async`.
-        let backend = NatsSessionLogBackend::new(self.jetstream.clone(), &activation.session_id)
-            .with_after_seq_observer(Arc::clone(&after_seq_observer));
-
-        // Abort turns promptly if lease is lost.
-        let watch_task =
-            Self::spawn_lease_loss_watch(&lease, &abort_signal, &activation.session_id);
-
-        let result = harnx_core::sink::with_agent_event_sink(event_sink, async {
-            // A half-installed agent cannot render its prompt, so stop here and
-            // let the caller record the failure rather than failing deeper in.
-            agent_setup?;
-
-            // Activation high-water cursor: max log seq of ANY user message we've fed into this
-            // activation. Includes messages folded into turn inputs AND mid-round injections.
-            // A continuation turn only runs if there's a user message with seq > this cursor.
-            let mut activation_high_water: Option<u64> = None;
-
-            loop {
-                if !lease.is_held() {
-                    break Ok(());
-                }
-
-                let (input, seed_cursor) = self
-                    .derive_turn_input(
-                        &activation,
-                        &lease,
-                        &per_session,
-                        &backend,
-                        self.session_index.as_ref(),
-                        activation_high_water,
-                    )
-                    .await;
-
-                log::info!(
-                    "execute_session turn: session_id={} seed_cursor={:?}",
-                    activation.session_id,
-                    seed_cursor,
-                );
-
-                // Initialize cursor for this turn. Use seed_cursor (from derive_turn_input)
-                // which is the max seq of messages folded into the input.
-                let turn_cursor = Arc::new(AtomicU64::new(seed_cursor.unwrap_or(0)));
-
-                // Update activation high-water if seed_cursor is higher.
-                if let Some(seed) = seed_cursor {
-                    activation_high_water = Some(activation_high_water.map_or(seed, |h| h.max(seed)));
-                }
-
-                let on_tool_round =
-                    build_mid_turn_injection_callback(backend.clone(), Arc::clone(&turn_cursor));
-
-                // Per-turn observer for the S2 header-insert migration seq. Only
-                // the first activation of a headerless session migrates; the
-                // migration re-maps the leading-user block onto this seq, which
-                // the turn answers. We must advance the activation high-water
-                // past it so the end-of-turn drain does not re-fold the remapped
-                // (already-answered) users and spuriously re-run the turn (S3).
-                let header_insert_seq = Arc::new(AtomicU64::new(0));
-
-                run_agent_loop_with_nats_inner(
-                    RunAgentLoopArgs {
-                        cluster_key: &self.cluster,
-                        manage_servers: self.manage_servers,
-                        session_id: &activation.session_id,
-                        config: per_session.clone(),
-                        instance_id: self.instance_id.clone(),
-                        initial_input: input,
-                        abort_signal: abort_signal.clone(),
-                        call_fn: self.call_fn.clone(),
-                        lease: None,
-                        lease_config: self.lease.clone(),
-                        after_seq_observer: None,
-                        header_insert_observer: None,
-                        session_index: self.session_index.as_ref(),
-                        on_tool_round: Some(on_tool_round),
-                        working_dir: None,
-                    }
-                    .with_lease(Arc::clone(&lease))
-                    .with_after_seq_observer(Arc::clone(&after_seq_observer))
-                    .with_header_insert_observer(Arc::clone(&header_insert_seq)),
-                )
-                .await?;
-
-                // After turn completes, update activation high-water from turn_cursor.
-                // turn_cursor was updated by mid-round injection callback for any messages
-                // injected during multi-round tool execution.
-                let turn_cursor_val = turn_cursor.load(Ordering::SeqCst);
-                if turn_cursor_val > 0 {
-                    activation_high_water = Some(activation_high_water.map_or(turn_cursor_val, |h| h.max(turn_cursor_val)));
-                }
-
-                // Advance past the header-insert migration seq (if any). The
-                // migration remaps this turn's leading-user block onto this seq,
-                // so treat it as consumed by this turn's answer.
-                let header_insert_val = header_insert_seq.load(Ordering::SeqCst);
-                if header_insert_val > 0 {
-                    activation_high_water = Some(
-                        activation_high_water.map_or(header_insert_val, |h| h.max(header_insert_val)),
-                    );
-                }
-
-                log::info!(
-                    "execute_session turn complete: session_id={} turn_cursor={} activation_high_water={:?}",
-                    activation.session_id,
-                    turn_cursor_val,
-                    activation_high_water,
-                );
-
-                if !lease.is_held() {
-                    break Ok(());
-                }
-
-                // DRAIN DECISION: cursor-based, not barrier-based.
-                // Re-run another turn ONLY if there's a user message with seq > activation_high_water.
-                // This prevents re-running when we've already consumed everything.
-                // Use the fresh leader-authoritative load so this re-read reflects
-                // both the worker's own just-persisted turn barrier and any client
-                // edit/retract committed just before the read (otherwise it would
-                // re-fold already-answered or retracted messages).
-                let tail = backend.load_events_latest_async().await?;
-
-                // Check for resumable in-flight tool rounds (multi-turn tool execution).
-                // Use reconstruct_state_from_nats to preserve NATS seqs for EditEntries resolution.
-                let reconstructed = harnx_core::session_reconstruct::reconstruct_state_from_nats(&tail);
-                let has_resumable = reconstructed.resumable_ctx.is_some();
-
-                // Check for new user messages beyond the high-water cursor.
-                let (new_messages, latest_new_seq) = fold_new_user_messages_since(&tail, activation_high_water);
-
-                log::info!(
-                    "execute_session drain check: session_id={} new_messages_count={} has_resumable={} latest_new_seq={:?}",
-                    activation.session_id,
-                    new_messages.len(),
-                    has_resumable,
-                    latest_new_seq,
-                );
-
-                // Continue only if there are genuinely new user messages OR a resumable tool context.
-                // A completed turn with nothing new => exactly one execution.
-                //
-                // Do NOT advance `activation_high_water` here. The drain only
-                // DETECTS that unanswered messages exist; the continuation turn
-                // CONSUMES them and advances the high-water from its own
-                // `seed_cursor` at the top of the loop. Advancing here would mark
-                // the messages consumed before the turn runs, so the continuation
-                // turn would derive an empty input and never answer them.
-                if new_messages.is_empty() && !has_resumable {
-                    break Ok(());
-                }
-            }
-        })
-        .await;
-
-        // Record the failure durably BEFORE releasing the lease: attached
-        // clients treat an `Error` entry as the turn barrier, and a client that
-        // reconnects later still sees why the turn produced nothing.
-        if let Err(error) = &result {
-            Self::record_session_error(&backend, &lease, error).await;
-        }
-
-        if !lease.is_held() {
-            log::warn!(
-                "session execution ended after failover: session_id={} worker_id={} revision={}",
-                activation.session_id,
-                lease.worker_id(),
-                lease.fence_token()
-            );
-        }
-
-        watch_task.abort();
-        control_task.abort();
-        let _ = tokio::time::timeout(std::time::Duration::from_millis(100), watch_task).await;
-        let _ = tokio::time::timeout(std::time::Duration::from_millis(100), control_task).await;
-
-        let _ = lease.release().await;
-        result
-    }
-
-    /// Append an `Error` entry for a turn that failed.
-    ///
-    /// Skipped when the lease is gone: a newer worker owns the session and
-    /// writing behind it would corrupt the log. That case is covered by the
-    /// client's orphan watchdog instead.
-    async fn record_session_error(
-        backend: &NatsSessionLogBackend,
-        lease: &NatsSessionLease,
-        error: &anyhow::Error,
-    ) {
-        if !should_append_control_log_entry(lease) {
-            return;
-        }
-        let entry = harnx_core::session::SessionLogEntry::Error {
-            message: format!("{error:#}"),
-            fence_token: lease.fence_token(),
-            timestamp: Some(chrono::Utc::now()),
-        };
-        if let Err(append_error) = backend.append_event(&entry).await {
-            log::warn!(
-                "failed to append Error entry: session_id={} err={append_error:#}",
-                backend.session_id(),
-            );
-        }
-    }
-
-    /// Spawn a task that watches for lease loss and aborts on loss.
-    fn spawn_lease_loss_watch(
-        lease: &Arc<NatsSessionLease>,
-        abort_signal: &crate::utils::AbortSignal,
-        session_id: &str,
-    ) -> tokio::task::JoinHandle<()> {
-        let mut lost = lease.lost_watch();
-        let abort_for_watch = abort_signal.clone();
-        let watch_session_id = session_id.to_string();
-        let watch_lease = Arc::clone(lease);
-        tokio::spawn(async move {
-            while lost.changed().await.is_ok() {
-                if !*lost.borrow() {
-                    log::warn!(
-                        "failover abort: session_id={} worker_id={} revision={} reason=lease_lost",
-                        watch_session_id,
-                        watch_lease.worker_id(),
-                        watch_lease.fence_token()
-                    );
-                    abort_for_watch.set_ctrlc();
-                    break;
-                }
-            }
-        })
-    }
-
-    /// Subscribe to control before spawning its listener task.
-    ///
-    /// Returning only after `subscribe` completes is the ordering barrier used by
-    /// activation handling before it acknowledges the non-durable work message.
-    async fn spawn_control_listener(ctx: ControlListenerCtx<'_>) -> Result<JoinHandle<()>> {
-        let ctrl_subject = control_subject(ctx.session_id);
-        let subscriber = ctx
-            .client
-            .subscribe(ctrl_subject)
-            .await
-            .context("subscribe to session control subject")?;
-        // Flush the SUB protocol command so broker-side interest exists before
-        // activation is acknowledged and clients can observe that readiness.
-        ctx.client
-            .flush()
-            .await
-            .context("flush session control subscription")?;
-        let ctrl_abort = ctx.abort_signal.clone();
-        let ctrl_lease = Arc::clone(ctx.lease);
-        let ctrl_backend = ctx.backend.clone();
-        Ok(tokio::spawn(async move {
-            use futures_util::StreamExt;
-            let mut messages = subscriber;
-            while let Some(msg) = messages.next().await {
-                // Parse the control command.
-                let Ok(command) = ControlCommand::from_bytes(&msg.payload) else {
-                    log::debug!("invalid control command payload, ignoring");
-                    continue;
-                };
-                match command {
-                    ControlCommand::Cancel => {
-                        // Worker-originated: append Cancel entry (worker's fence token),
-                        // THEN fire AbortSignal.
-                        if should_append_control_log_entry(&ctrl_lease) {
-                            let cancel_entry = harnx_core::session::SessionLogEntry::Cancel {
-                                fence_token: ctrl_lease.fence_token(),
-                            };
-                            // Append BEFORE firing abort; if append fails, still abort (safe).
-                            if let Err(e) = ctrl_backend.append_event_blocking(&cancel_entry) {
-                                log::warn!("failed to append Cancel entry: {e}");
-                            }
-                        }
-                        ctrl_abort.set_ctrlc();
-                    }
-                }
-            }
-        }))
-    }
-
-    /// Reconstruct session state using the canonical algorithm.
-    ///
-    /// Returns the session's turn status, effective pending message, and
-    /// resumable context for driving the agent loop correctly.
-    async fn reconstruct_session_state(
-        &self,
-        backend: &NatsSessionLogBackend,
-    ) -> harnx_core::session_reconstruct::ReconstructedState {
-        match backend.load_events_latest_async().await {
-            Ok(entries) => harnx_core::session_reconstruct::reconstruct_state_from_nats(&entries),
-            Err(err) => {
-                log::warn!(
-                    "failed to load session log for reconstruction: session_id={} worker_id={} err={err}",
-                    backend.session_id(),
-                    self.worker_id,
-                );
-                harnx_core::session_reconstruct::ReconstructedState {
-                    turn_status: harnx_core::session_reconstruct::TurnStatus::Idle,
-                    next_turn_messages: Vec::new(),
-                    resumable_ctx: None,
-                }
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        configured_worker_services, optional_tool_server, should_start_tool_servers,
-        tool_servers_matching_use_tools,
-    };
-    use crate::config::ToolServerConfig;
-
-    fn tool_server(name: &str) -> ToolServerConfig {
-        ToolServerConfig {
-            name: name.to_string(),
-            command: format!("harnx-{name}-server"),
-            args: Vec::new(),
-            env: Default::default(),
-            enabled: true,
-            description: None,
-            package: None,
-            hooks: None,
-        }
-    }
-
-    #[test]
-    fn should_start_tool_servers_requires_managing_and_a_nonempty_list() {
-        let servers = vec![tool_server("time")];
-        assert!(
-            should_start_tool_servers(true, &servers),
-            "managing with servers configured must start them"
-        );
-        assert!(
-            !should_start_tool_servers(true, &[]),
-            "managing with nothing configured must not resolve a broker to spawn nothing"
-        );
-        assert!(
-            !should_start_tool_servers(false, &servers),
-            "a consuming worker must not spawn its own tool servers"
-        );
-    }
-
-    #[test]
-    fn configured_worker_services_falls_back_to_config_use_tools_without_agent() {
-        let config = crate::config::GlobalConfig::default();
-        {
-            let mut config = config.write();
-            config.agent = None;
-            config.use_tools = Some(vec!["time_get_current_time".to_string()]);
-            config.tool_servers = vec![tool_server("time"), tool_server("weather")];
-        }
-
-        let (servers, _hooks) = configured_worker_services(&config);
-
-        assert_eq!(servers.len(), 1);
-        assert_eq!(servers[0].name, "time");
-    }
-
-    #[test]
-    fn tool_server_filter_matches_selectors_wildcard_and_absent_use_tools() {
-        let servers = [tool_server("time"), tool_server("weather")];
-
-        let matching =
-            tool_servers_matching_use_tools(&servers, None, &["time_get_current_time".to_string()]);
-        assert_eq!(
-            matching
-                .iter()
-                .map(|server| server.name.as_str())
-                .collect::<Vec<_>>(),
-            ["time"]
-        );
-        assert_eq!(
-            tool_servers_matching_use_tools(&servers, None, &["*".to_string()]).len(),
-            2
-        );
-        assert!(tool_servers_matching_use_tools(&servers, None, &[]).is_empty());
-    }
-
-    #[test]
-    fn tool_server_filter_uses_package_scoped_display_name() {
-        let mut time = tool_server("time");
-        time.package = Some("other/tools".to_string());
-        let servers = [time, tool_server("weather")];
-
-        let matching = tool_servers_matching_use_tools(
-            &servers,
-            Some("active"),
-            &["other__tools__time_get_current_time".to_string()],
-        );
-        assert_eq!(matching.len(), 1);
-        assert_eq!(matching[0].name, "time");
-    }
-
-    #[test]
-    fn tool_server_filter_deduplicates_full_identities_first_wins() {
-        let user_time = tool_server("time");
-        let mut package_time = tool_server("time");
-        package_time.package = Some("coding".to_string());
-        let weather = tool_server("weather");
-        let servers = [user_time, package_time, weather];
-
-        let matching = tool_servers_matching_use_tools(&servers, None, &["*".to_string()]);
-
-        assert_eq!(matching.len(), 3);
-        assert_eq!(matching[0].name, "time");
-        assert_eq!(matching[0].package, None);
-        assert_eq!(matching[1].package.as_deref(), Some("coding"));
-        assert_eq!(matching[2].name, "weather");
-
-        let mut aaa = tool_server("dup");
-        aaa.package = Some("aaa".to_string());
-        let mut zzz = tool_server("dup");
-        zzz.package = Some("zzz".to_string());
-        let package_matching =
-            tool_servers_matching_use_tools(&[aaa, zzz], None, &["*".to_string()]);
-        assert_eq!(package_matching.len(), 2);
-        assert_eq!(package_matching[0].package.as_deref(), Some("aaa"));
-        assert_eq!(package_matching[1].package.as_deref(), Some("zzz"));
-
-        let duplicate = tool_server("same");
-        let duplicate_matching = tool_servers_matching_use_tools(
-            &[duplicate.clone(), duplicate],
-            None,
-            &["*".to_string()],
-        );
-        assert_eq!(duplicate_matching.len(), 1);
-    }
-
-    #[test]
-    fn worker_startup_continues_when_configured_binary_is_missing() {
-        harnx_core::require_nextest();
-        let missing_dir = tempfile::tempdir().expect("create missing-binary test directory");
-        let mut missing_server = tool_server("time");
-        missing_server.command = missing_dir
-            .path()
-            .join("harnx-missing-tool-server")
-            .to_string_lossy()
-            .into_owned();
-        let filtered = tool_servers_matching_use_tools(
-            &[missing_server],
-            None,
-            &["time_get_current_time".to_string()],
-        );
-
-        assert_eq!(filtered.len(), 1);
-        // Supervisor's missing-binary path is soft-fail and therefore reaches
-        // daemon startup as Ok; integration coverage asserts its warning.
-        assert!(optional_tool_server::<()>(Ok(())).is_some());
-    }
 }
